[FIX] Renamed main package from cuetojson to dagviz
